requirements.txt
# Core dependencies
openai>=1.0.0
python-dotenv>=1.0.0
pyyaml>=6.0

# MCP SDK for tools
mcp>=1.9.0

# Web server dependencies
flask>=3.0.0
flask-socketio>=5.3.0
flask-cors>=4.0.0

# Additional dependencies
requests>=2.31.0
urllib3>=2.0.0
beautifulsoup4>=4.12.0
aiohttp>=3.9.0

# LaTeX and document processing
# Note: You need to install LaTeX distribution separately (TeX Live, MiKTeX, etc.)
# This tool requires pdflatex, xelatex, or lualatex to be available in PATH

# Development dependencies (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0
# black>=23.0.0
# flake8>=6.0.0 